refactor: delete batch_request_anonymizer and checkpoint_anonymizer
